Test where a date falls relative to a date range

Test where a date falls relative to a date range

date_in_range(target, start, end, .before = FALSE, .ongoing = TRUE,
  .during = TRUE, .after = FALSE)

Arguments

target
The date of interest

start
The starting date of the range

end
The ending date of the range

.before
The value to assign if the date falls before the range (default FALSE)

.ongoing
The value to assign if the date falls after start, and end is NA (default TRUE)

.during
The value to assign if the date falls within the range (default TRUE)

.after
The value to assign if the date falls after the range (default FALSE)

Value

The return type is determined by the user. The default is logical.

Details

The return values can be logical, character, or numeric. The interval for within includes the start and end dates.

Examples

date_in_range(iso_to_posix("2018-03-14T12:34"), iso_to_posix("2018-03-11T12:34"), iso_to_posix("2018-03-18T12:34"))

#> [1] TRUE

date_in_range("2018-03-14", "2018-03", "2018-03-18T12:34")

#> [1] TRUE

date_in_range(iso_to_posix("2018-03-14T12:34"), iso_to_posix("2018-03-15T12:34"), iso_to_posix("2018-03-18T12:34"))

#> [1] FALSE

date_in_range(iso_to_posix("2018-03-14T12:34")+(1:10)*24*2600,
  iso_to_posix("2018-03-11T12:34"),
  iso_to_posix("2018-03-18T12:34"))

#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE

date_in_range(iso_to_posix("2018-03-09T12:34")+(1:10)*24*3600,
  iso_to_posix("2018-03-11T12:34"),
  c(rep(iso_to_posix("2018-03-15T12:34"),9), NA),
  .before="BEFORE", .ongoing="ONGOING", .during="DURING", .after="AFTER")

#>  [1] "BEFORE"  "DURING"  "DURING"  "DURING"  "DURING"  "DURING"  "AFTER"  
#>  [8] "AFTER"   "AFTER"   "ONGOING"