Create a default value in a dataframe column if the column doesn't already exist

Create a default value in a dataframe column if the column doesn't already exist

default_missing_column(column, default_value = NA, env = dynGet("."))

Arguments

column
string name for column to default

default_value
default value for column, of length 1 or number of rows of dataframe

env
the environment (dataframe) in which to evaluate the column

Value

the values of either the existing column, or the default_value if the column does not exist

Details

This function is designed to work from within a dplyr mutate call.

Examples

library(dplyr)
mtcars %>% as_data_frame() %>% mutate(newcol=default_missing_column("newcol", 1))

#> Warning: `as_data_frame()` is deprecated, use `as_tibble()` (but mind the new semantics).
#> This warning is displayed once per session.

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb newcol
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4      1
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4      1
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1      1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1      1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2      1
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1      1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4      1
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2      1
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2      1
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4      1
#> # … with 22 more rows

mtcars %>% as_data_frame() %>% mutate(newcol=default_missing_column("newcol", 1:32))

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb newcol
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <int>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4      1
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4      2
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1      3
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1      4
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2      5
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1      6
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4      7
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2      8
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2      9
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4     10
#> # … with 22 more rows