Processes domain data for specified dependent variables (DV)

Processes domain data for specified dependent variables (DV)

getDV(.data, ID, TIME, DV, BQL, LLOQ, EVID, MDV, dv.filter, Units, ...)

Arguments

.data
The domain data frame

ID
A column name or expression for the subject identifier.

TIME
A column name or expression for the TIME column.

DV
A column name or expression for the dependent variable (DV).

BQL
An expression returning TRUE when the observation is BQL.

LLOQ
An expression giving the LLOQ value to used in replacing DV when BQL is TRUE.

EVID
A column name or expression for the EVID column (defaults to 0).

MDV
A column name or expression for the MDV column (defaults to 0).

dv.filter
Filter expression to apply to subset the data.

Units
A column name or expression containing the units of DV.

...
Additional column mappings.

Value

A data_frame with the mapped data

Details

The input data is filtered, if a filter expression is provided. ID, TIME, and DV are required. If BQL and LLOQ are supplied, then DV will be set to the LLOQ value when BQL is TRUE. EVID can be provided as a number or expression, but will default to 0 if not supplied. Additional column mappings will be added to the returned data.

This function uses non-standard evaluation, so the arguments do not need to be quoted. For example, to compute a date from ISO-8601 strings include TIME=parsedate::parse_date(EXSTDTC). Of course the actual column may have a different name. Extra arguments (...) must be named, for example: MYVALUE = pmax(0,DV)