Processes domain data for individual doses

Processes domain data for individual doses

getIndividualDoses(.data, ID, TIME, AMT, EVID, MDV, ex.filter, Units,
  combine.tol = 0, ...)

Arguments

.data
The domain data frame

ID
A column name or expression for the subject identifier.

TIME
A column name or expression for the TIME column.

AMT
A column name or expression for the AMT column.

EVID
A column name or expression for the EVID column (defaults to 1).

MDV
A column name or expression for the MDV column (defaults to 1).

ex.filter
Filter expression to apply to subset the data.

Units
A column name or expression containing the units of AMT.

...
Additional column mappings.

Value

A data_frame with the mapped data

Details

The input data is filtered, if a filter expression is provided. ID, TIME, and AMT are required. EVID can be provided as a number or expression, but will default to 1 if not supplied. Additional column mappings will be added to the returned data.

This function uses non-standard evaluation, so the arguments do not need to be quoted. For example, to compute a date from ISO-8601 strings include TIME=parsedate::parse_date(EXSTDTC). Of course the actual column may have a different name. Extra arguments (...) must be named, for example: MYVALUE = pmax(0,DV).

AMT units may be supplied as a column reference (eg, EXDOSU) or as a fixed string (eg, "mg"). If only a single unit is represented then AMT will be converted to a pmunits column, where the units are stored as a column attribute. Otherweise a column AMTU will be created and carried through to the output.