Find a baseline value

Find a baseline value

get_baseline(col, .filter, ..., fun.summ = mean_)

Arguments

col
An unquoted variable name

.filter
An unquoted R logical expression (evaluates to TRUE or FALSE). Pass TRUE to select all records for summarization.

...
Comma separated, unquoted, additional grouping columns

fun.summ
An unquoted function name to use to summarise multiple baseline values per group

Value

Returns a vector of selected baseline values

Details

get_baseline can be used inside a mutate or summarise call, and therefore inside the get mapped value functions (e.g., getDV). It finds the rows that satisfy the criteria in the .filter argument, then applies fun.summ (a summary function) to those rows. In this way, multiple values can be used for a baseline value.

Examples

library(dplyr)
library(PMDatR)
Theoph %>% mutate(conc.bl = get_baseline(conc, Time>0 & Time<1, Subject, fun.summ=mean_))

#>     Subject   Wt Dose  Time  conc conc.bl
#> 1         1 79.6 4.02  0.00  0.74   4.705
#> 2         1 79.6 4.02  0.25  2.84   4.705
#> 3         1 79.6 4.02  0.57  6.57   4.705
#> 4         1 79.6 4.02  1.12 10.50   4.705
#> 5         1 79.6 4.02  2.02  9.66   4.705
#> 6         1 79.6 4.02  3.82  8.58   4.705
#> 7         1 79.6 4.02  5.10  8.36   4.705
#> 8         1 79.6 4.02  7.03  7.47   4.705
#> 9         1 79.6 4.02  9.05  6.89   4.705
#> 10        1 79.6 4.02 12.12  5.94   4.705
#> 11        1 79.6 4.02 24.37  3.28   4.705
#> 12        2 72.4 4.40  0.00  0.00   4.815
#> 13        2 72.4 4.40  0.27  1.72   4.815
#> 14        2 72.4 4.40  0.52  7.91   4.815
#> 15        2 72.4 4.40  1.00  8.31   4.815
#> 16        2 72.4 4.40  1.92  8.33   4.815
#> 17        2 72.4 4.40  3.50  6.85   4.815
#> 18        2 72.4 4.40  5.02  6.08   4.815
#> 19        2 72.4 4.40  7.03  5.40   4.815
#> 20        2 72.4 4.40  9.00  4.55   4.815
#> 21        2 72.4 4.40 12.00  3.01   4.815
#> 22        2 72.4 4.40 24.30  0.90   4.815
#> 23        3 70.5 4.53  0.00  0.00   5.650
#> 24        3 70.5 4.53  0.27  4.40   5.650
#> 25        3 70.5 4.53  0.58  6.90   5.650
#> 26        3 70.5 4.53  1.02  8.20   5.650
#> 27        3 70.5 4.53  2.02  7.80   5.650
#> 28        3 70.5 4.53  3.62  7.50   5.650
#> 29        3 70.5 4.53  5.08  6.20   5.650
#> 30        3 70.5 4.53  7.07  5.30   5.650
#> 31        3 70.5 4.53  9.00  4.90   5.650
#> 32        3 70.5 4.53 12.15  3.70   5.650
#> 33        3 70.5 4.53 24.17  1.05   5.650
#> 34        4 72.7 4.40  0.00  0.00   3.245
#> 35        4 72.7 4.40  0.35  1.89   3.245
#> 36        4 72.7 4.40  0.60  4.60   3.245
#> 37        4 72.7 4.40  1.07  8.60   3.245
#> 38        4 72.7 4.40  2.13  8.38   3.245
#> 39        4 72.7 4.40  3.50  7.54   3.245
#> 40        4 72.7 4.40  5.02  6.88   3.245
#> 41        4 72.7 4.40  7.02  5.78   3.245
#> 42        4 72.7 4.40  9.02  5.33   3.245
#> 43        4 72.7 4.40 11.98  4.19   3.245
#> 44        4 72.7 4.40 24.65  1.15   3.245
#> 45        5 54.6 5.86  0.00  0.00   3.825
#> 46        5 54.6 5.86  0.30  2.02   3.825
#> 47        5 54.6 5.86  0.52  5.63   3.825
#> 48        5 54.6 5.86  1.00 11.40   3.825
#> 49        5 54.6 5.86  2.02  9.33   3.825
#> 50        5 54.6 5.86  3.50  8.74   3.825
#> 51        5 54.6 5.86  5.02  7.56   3.825
#> 52        5 54.6 5.86  7.02  7.09   3.825
#> 53        5 54.6 5.86  9.10  5.90   3.825
#> 54        5 54.6 5.86 12.00  4.37   3.825
#> 55        5 54.6 5.86 24.35  1.57   3.825
#> 56        6 80.0 4.00  0.00  0.00   2.185
#> 57        6 80.0 4.00  0.27  1.29   2.185
#> 58        6 80.0 4.00  0.58  3.08   2.185
#> 59        6 80.0 4.00  1.15  6.44   2.185
#> 60        6 80.0 4.00  2.03  6.32   2.185
#> 61        6 80.0 4.00  3.57  5.53   2.185
#> 62        6 80.0 4.00  5.00  4.94   2.185
#> 63        6 80.0 4.00  7.00  4.02   2.185
#> 64        6 80.0 4.00  9.22  3.46   2.185
#> 65        6 80.0 4.00 12.10  2.78   2.185
#> 66        6 80.0 4.00 23.85  0.92   2.185
#> 67        7 64.6 4.95  0.00  0.15   1.600
#> 68        7 64.6 4.95  0.25  0.85   1.600
#> 69        7 64.6 4.95  0.50  2.35   1.600
#> 70        7 64.6 4.95  1.02  5.02   1.600
#> 71        7 64.6 4.95  2.02  6.58   1.600
#> 72        7 64.6 4.95  3.48  7.09   1.600
#> 73        7 64.6 4.95  5.00  6.66   1.600
#> 74        7 64.6 4.95  6.98  5.25   1.600
#> 75        7 64.6 4.95  9.00  4.39   1.600
#> 76        7 64.6 4.95 12.05  3.53   1.600
#> 77        7 64.6 4.95 24.22  1.15   1.600
#> 78        8 70.5 4.53  0.00  0.00   4.470
#> 79        8 70.5 4.53  0.25  3.05   4.470
#> 80        8 70.5 4.53  0.52  3.05   4.470
#> 81        8 70.5 4.53  0.98  7.31   4.470
#> 82        8 70.5 4.53  2.02  7.56   4.470
#> 83        8 70.5 4.53  3.53  6.59   4.470
#> 84        8 70.5 4.53  5.05  5.88   4.470
#> 85        8 70.5 4.53  7.15  4.73   4.470
#> 86        8 70.5 4.53  9.07  4.57   4.470
#> 87        8 70.5 4.53 12.10  3.00   4.470
#> 88        8 70.5 4.53 24.12  1.25   4.470
#> 89        9 86.4 3.10  0.00  0.00   8.200
#> 90        9 86.4 3.10  0.30  7.37   8.200
#> 91        9 86.4 3.10  0.63  9.03   8.200
#> 92        9 86.4 3.10  1.05  7.14   8.200
#> 93        9 86.4 3.10  2.02  6.33   8.200
#> 94        9 86.4 3.10  3.53  5.66   8.200
#> 95        9 86.4 3.10  5.02  5.67   8.200
#> 96        9 86.4 3.10  7.17  4.24   8.200
#> 97        9 86.4 3.10  8.80  4.11   8.200
#> 98        9 86.4 3.10 11.60  3.16   8.200
#> 99        9 86.4 3.10 24.43  1.12   8.200
#> 100      10 58.2 5.50  0.00  0.24   4.055
#> 101      10 58.2 5.50  0.37  2.89   4.055
#> 102      10 58.2 5.50  0.77  5.22   4.055
#> 103      10 58.2 5.50  1.02  6.41   4.055
#> 104      10 58.2 5.50  2.05  7.83   4.055
#> 105      10 58.2 5.50  3.55 10.21   4.055
#> 106      10 58.2 5.50  5.05  9.18   4.055
#> 107      10 58.2 5.50  7.08  8.02   4.055
#> 108      10 58.2 5.50  9.38  7.14   4.055
#> 109      10 58.2 5.50 12.10  5.68   4.055
#> 110      10 58.2 5.50 23.70  2.42   4.055
#> 111      11 65.0 4.92  0.00  0.00   6.700
#> 112      11 65.0 4.92  0.25  4.86   6.700
#> 113      11 65.0 4.92  0.50  7.24   6.700
#> 114      11 65.0 4.92  0.98  8.00   6.700
#> 115      11 65.0 4.92  1.98  6.81   6.700
#> 116      11 65.0 4.92  3.60  5.87   6.700
#> 117      11 65.0 4.92  5.02  5.22   6.700
#> 118      11 65.0 4.92  7.03  4.45   6.700
#> 119      11 65.0 4.92  9.03  3.62   6.700
#> 120      11 65.0 4.92 12.12  2.69   6.700
#> 121      11 65.0 4.92 24.08  0.86   6.700
#> 122      12 60.5 5.30  0.00  0.00   2.605
#> 123      12 60.5 5.30  0.25  1.25   2.605
#> 124      12 60.5 5.30  0.50  3.96   2.605
#> 125      12 60.5 5.30  1.00  7.82   2.605
#> 126      12 60.5 5.30  2.00  9.72   2.605
#> 127      12 60.5 5.30  3.52  9.75   2.605
#> 128      12 60.5 5.30  5.07  8.57   2.605
#> 129      12 60.5 5.30  7.07  6.59   2.605
#> 130      12 60.5 5.30  9.03  6.11   2.605
#> 131      12 60.5 5.30 12.05  4.57   2.605
#> 132      12 60.5 5.30 24.15  1.17   2.605

Theoph %>% mutate(conc=set_units(conc,"mg/ml"),
  conc.bl = get_baseline(conc, Time>0 & Time<1, Subject, fun.summ=mean_)) %>% as_data_frame

#> # A tibble: 132 x 6
#>    Subject    Wt  Dose   Time conc    conc.bl
#>    <ord>   <dbl> <dbl>  <dbl> <mg/ml> <mg/ml>
#>  1 1        79.6  4.02  0     " 0.74" 4.705  
#>  2 1        79.6  4.02  0.25  " 2.84" 4.705  
#>  3 1        79.6  4.02  0.570 " 6.57" 4.705  
#>  4 1        79.6  4.02  1.12  10.50   4.705  
#>  5 1        79.6  4.02  2.02  " 9.66" 4.705  
#>  6 1        79.6  4.02  3.82  " 8.58" 4.705  
#>  7 1        79.6  4.02  5.1   " 8.36" 4.705  
#>  8 1        79.6  4.02  7.03  " 7.47" 4.705  
#>  9 1        79.6  4.02  9.05  " 6.89" 4.705  
#> 10 1        79.6  4.02 12.1   " 5.94" 4.705  
#> # ... with 122 more rows