Convert an ISO8601 date/time to POSIXct

Convert an ISO8601 date/time to POSIXct

iso_to_posix(dtc, .time, .tz = "UTC", .day, .month)

Arguments

dtc
A ISO8601 formatted date/time

.time
If dtc is date-only, a time to use as an offset. If length is not 1 or same as dtc, .time is recycled with a warning.

.tz
The timezone for the dtc values. Default is UTC.

.day
If dtc is year-month only, a month to use a default. If length is not 1 or same as dtc, .time is recycled with a warning.

.month
If dtc is year only, a month to use a default. If length is not 1 or same as dtc, .time is recycled with a warning.

Value

date/time as R POSIXct native type.

Details

The conversion does its best to guess at the format of the time and to fill in any missing parts, defaulting to Jan, the first of the month, and midnight. The default day, month, and time can be provided, but it's important to note that the infill of these values won't cross more than one division. That is, "1990" will default to "1990-01-01 00:00", even if .time is given as 06:35. The missing pieces must be filled in in order.

Examples

iso_to_posix("2016-10-03T21:22:45")

#> [1] "2016-10-03 21:22:45 UTC"

iso_to_posix(c("2016-10-03","2016-10-03T12:34"),"09:00")

#> [1] "2016-10-03 09:00:00 UTC" "2016-10-03 12:34:00 UTC"

iso_to_posix(c("2016-10-03","2016-10-03","2016-10-03"), c("09:00", "10:10")) #gives a warning about length of .times

#> Warning: .time should be equal in length to dtc or of length 1.  .time has been replicated to the required length.

#> [1] "2016-10-03 09:00:00 UTC" "2016-10-03 10:10:00 UTC"
#> [3] "2016-10-03 09:00:00 UTC"

iso_to_posix(c("1980", "1980-06", "1980-06-15"), .time="09:34")

#> [1] "1980-01-01 00:00:00 UTC" "1980-06-01 00:00:00 UTC"
#> [3] "1980-06-15 09:34:00 UTC"

iso_to_posix(c("1980", "1980-06", "1980-06-15"), .time="09:34", .day=15, .month=6)

#> [1] "1980-06-15 09:34:00 UTC" "1980-06-15 09:34:00 UTC"
#> [3] "1980-06-15 09:34:00 UTC"